csv tests to integration tests
