+++ polars/tests/it/main.rs
@@ -1,2 +1,3 @@
+mod io;
 mod joins;
 mod lazy;
